Move modal logic to sidebar; prefill edit modal with current title
